requirements.txt
certifi==2020.6.20
crispy_bootstrap5==0.6
Django~=3.1.12
django_bootstrap_modal_forms==2.2.0
django-bootstrap5==21.3
django-cleanup==4.0.0
django_crispy_forms==1.14.0
django_formtools==2.3
python_dateutil==2.8.2
requests==2.22.0
six==1.15.0
timeago==1.0.15
urllib3==1.25.7
